+++ src/main/java/com/rabobank/repository/CustomerStatementsRepository.java
@@ -1,13 +1,11 @@
 package com.rabobank.repository;
 
-import java.math.BigInteger;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.rabobank.domain.CustomerStatements;
 
 @Repository
-public interface CustomerStatementsRepository extends JpaRepository<CustomerStatements,BigInteger>{
+public interface CustomerStatementsRepository extends JpaRepository<CustomerStatements,Long>{
 
 }
